Move value widget to its own module
